requirements.txt
PyPDF2
qrcode
Pillow
PyMuPDF
